moves event detection to new modules space
